logger successfully integrated with node manager

+++ node-manager/main.py
@@ -12,19 +12,20 @@ from confluent_kafka import Consumer, Producer
 from decouple import config
 from fastapi import FastAPI
 from heartbeat_service import HeartbeatService
+from logger_utils import Logger
 from Messenger import Produce
 from pymongo import MongoClient
 from storage import downloadFile
 
 KAFKA_CONFIG_FILE = "kafka_setup_config.json"
 TOPIC = "topic_node_manager"
+SERVICE_NAME = "node_manager"
 mongokey = config("mongoKey")
 client = MongoClient(mongokey)
 db = client["platform"]
 producer = Produce()
+logger = Logger()
 
-
-SERVICE_NAME = "Node Manager"
 # Create a new instance of the HeartbeatService class
 heartbeat_service = HeartbeatService("topic_node_manager_health", SERVICE_NAME)
 
@@ -103,6 +104,8 @@ class Consume:
 
 
 def deploy_app(appname: str, appid: str, userid: str):
+    logger.log(SERVICE_NAME, 1, "Deploying App.....")
+
     collection = db.App
     collection.create_index("name", unique=True)
     user_collection = db.User
@@ -113,30 +116,72 @@ def deploy_app(appname: str, appid: str, userid: str):
     curr_user = user_collection.find_one({"_id": ObjectId(userid)})
 
     if not curr_app:
-        message = {"src": TOPIC, "status": 404, "msg": "Could not find the app"}
-        produce.push("topic-internal-api", "", json.dumps(message))
+        # message = {"src": TOPIC, "status": 404, "msg": "Could not find the app"}
+        # produce.push("topic-internal-api", "", json.dumps(message))
+        logger.log(SERVICE_NAME, 3, f"{appname} does not exist in our records")
 
         return {"success": 404, "err": "App does not found"}
 
     ip = get_ip()
 
-    status = downloadFile("apps", f"{appname}.zip", ".")
+    logger.log(SERVICE_NAME, 1, f"Starting the ZIP File download.....")
+    res = downloadFile("apps", f"{appname}.zip", ".")
+    if res["status"] == False:
+        logger.log(SERVICE_NAME, 3, res["message"])
+        return {"success": 500, "err": "Could not download the file, please check logs"}
+
+    logger.log(SERVICE_NAME, 1, f"{appname}.zip downloaded successfully")
 
     with zipfile.ZipFile(f"{appname}.zip", "r") as zip_ref:
         zip_ref.extractall(".")
 
+    logger.log(SERVICE_NAME, 1, "Allocating resouces....")
+
     cmd = f"docker stop {appname} && docker rm {appname}"
-    os.system(cmd)
+    res = os.system(cmd)
+    if res != 0:
+        logger.log(
+            SERVICE_NAME,
+            4,
+            f"{res} error code occurs from the VM Machine, please check",
+        )
+        return {"success": False, "message": "Internal Server Error"}
     cmd = f"docker rmi {appname}"
-    os.system(cmd)
+    res = os.system(cmd)
+    if res != 0:
+        logger.log(
+            SERVICE_NAME,
+            4,
+            f"{res} error code occurs from the VM Machine, please check",
+        )
+        return {"success": False, "message": "Internal Server Error"}
     generate_docker_image(appname)
     cmd = f"docker build -t {appname} {appname}"
-    os.system(cmd)
+    res = os.system(cmd)
+    if res != 0:
+        logger.log(
+            SERVICE_NAME,
+            4,
+            f"{res} error code occurs from the VM Machine, please check",
+        )
+        return {"success": False, "message": "Internal Server Error"}
     assign_port = get_free_port()
     cmd = f"docker run --name {appname} -d -p {assign_port}:80 {appname}"
-    os.system(cmd)
+    res = os.system(cmd)
+    if res != 0:
+        logger.log(
+            SERVICE_NAME,
+            4,
+            f"{res} error code occurs from the VM Machine, please check",
+        )
+        return {"success": False, "message": "Internal Server Error"}
 
     data = {"active": True, "port": assign_port, "ip": ip}
+
+    logger.log(
+        SERVICE_NAME, 1, "App deployed successfully", app_name=appname, user_id=userid
+    )
+    logger.log(SERVICE_NAME, 1, json.dumps(data), app_name=appname, user_id=userid)
 
     collection.find_one_and_update({"name": appname}, {"$set": data})
 
@@ -158,12 +203,31 @@ def stop_app(appname: str, appid: str, userid: str):
     active = collection.find_one({"_id": ObjectId(appid), "user": ObjectId(userid)})
 
     if not active:
+        logger.log(
+            SERVICE_NAME, 3, "App does not exist", app_name=appname, user_id=userid
+        )
         return {"status": 404, "msg": "Could not find the app"}
     cmd = f"docker stop {appname}"
-    os.system(cmd)
+    res = os.system(cmd)
+    if res != 0:
+        logger.log(
+            SERVICE_NAME,
+            4,
+            f"{res} error code occurs from the VM Machine, please check",
+        )
+        return {"success": False, "message": "Internal Server Error"}
     data = {"active": False}
     collection.find_one_and_update(
         {"_id": ObjectId(appid), "user": ObjectId(userid)}, {"$set": data}
+    )
+
+    logger.log(SERVICE_NAME, 1, f"{appname} is stopped successfully")
+    logger.log(
+        SERVICE_NAME,
+        1,
+        f"{appname} is stopped successfully",
+        app_name=appname,
+        user_id=userid,
     )
     return data
 
@@ -173,13 +237,32 @@ def start_app(appname: str, appid: str, userid: str):
     active = collection.find_one({"_id": ObjectId(appid), "user": ObjectId(userid)})
 
     if not active:
+        logger.log(
+            SERVICE_NAME, 3, "App does not exist", app_name=appname, user_id=userid
+        )
         return {"status": 404, "msg": "Could not find the app"}
 
     cmd = f"docker start {appname}"
-    os.system(cmd)
+    res = os.system(cmd)
+    if res != 0:
+        logger.log(
+            SERVICE_NAME,
+            4,
+            f"{res} error code occurs from the VM Machine, please check",
+        )
+        return {"success": False, "message": "Internal Server Error"}
     data = {"active": True}
     collection.find_one_and_update(
         {"_id": ObjectId(appid), "user": ObjectId(userid)}, {"$set": data}
+    )
+
+    logger.log(SERVICE_NAME, 1, f"{appname} is stopped successfully")
+    logger.log(
+        SERVICE_NAME,
+        1,
+        f"{appname} is stopped successfully",
+        app_name=appname,
+        user_id=userid,
     )
     return data
 
@@ -226,9 +309,17 @@ def initialize():
         # upservices[service] = {"port": assign_port, "ip": ip}
         # data = {"name": service, "port": assign_port, "ip": ip, "active": True}
         # collection.insert_one(data)
+        logger.log(SERVICE_NAME, 1, f"{service}: STARTING....")
         cmd = f"kubectl apply -f ./{service}/manifests"
-        os.system(cmd)
+        res = os.system(cmd)
+        if res != 0:
+            logger.log(
+                SERVICE_NAME,
+                4,
+                f"{res} error code occurs from the VM Machine, please check",
+            )
 
+    logger.log(SERVICE_NAME, 1, "All the services have initialised")
     return {"services": upservices}
 
 
@@ -254,8 +345,17 @@ def destroy():
         # upservices[service] = {"port": assign_port, "ip": ip}
         # data = {"name": service, "port": assign_port, "ip": ip, "active": True}
         # collection.insert_one(data)
+        logger.log(SERVICE_NAME, 1, f"{service}: STOPPING....")
         cmd = f"kubectl delete -f ./{service}/manifests"
-        os.system(cmd)
+        res = os.system(cmd)
+        if res != 0:
+            logger.log(
+                SERVICE_NAME,
+                4,
+                f"{res} error code occurs from the VM Machine, please check",
+            )
+
+    logger.log(SERVICE_NAME, 1, "All the services have been stopped")
 
     return {"services": upservices}
